codereview comments incorporated

+++ src/ai/defaultai.h
@@ -329,11 +329,11 @@ private:
 	int32_t spots_;  // sum of buildable fields
 	int32_t vacant_mil_positions_;  // sum of vacant positions in militarysites and training sites
 	// statistics for training sites per type
-	uint8_t ts_basic_count_;
-	uint8_t ts_basic_const_count_;
-	uint8_t ts_advanced_count_;
-	uint8_t ts_advanced_const_count_;
-	uint8_t ts_without_trainers_;
+	int16_t ts_basic_count_;
+	int16_t ts_basic_const_count_;
+	int16_t ts_advanced_count_;
+	int16_t ts_advanced_const_count_;
+	int16_t ts_without_trainers_;
 
 	// This stores highest priority for new buildings except for militarysites
 	int32_t highest_nonmil_prio_;
